requirements.txt
fastapi==0.85.1
pyBKT==1.4
Pyrebase4==4.5.0
gunicorn==20.1.0
uvicorn==0.18.3
redis==4.3.4
python-multipart==0.0.5
tortoise-orm==0.19.2
passlib==1.7.4
PyJWT==2.6.0
asyncpg==0.27.0